Rename: Move Home, LoginPage, RegisterPage to src/pages/
