+++ src/main.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Home from '@components/Home.tsx';
+import Home from '@pages/Home.tsx';
+import LoginPage from '@pages/LoginPage.tsx';
+import RegisterPage from '@pages/RegisterPage.tsx';
 
 import '@css/index.css';
 
 import {rootStore} from '@/modules';
 import {Provider} from 'react-redux';
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
-import LoginPage from '@components/LoginPage.tsx';
-import RegisterPage from '@components/RegisterPage.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
